Refactoring: Moved responsibilities from minitwit.go to appropriate files
