+++ src/models/user.go
@@ -1,0 +1,8 @@
+package models
+
+type User struct {
+	UserID   int
+	Username string
+	Email    string
+	PwHash   string
+}
